ai-model-api/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
numpy==2.3.2
PyJWT==2.10.1
scikit-learn==1.7.1
scipy==1.16.1
SQLAlchemy==2.0.43
threadpoolctl==3.6.0
typing_extensions==4.15.0
Werkzeug==3.1.3
